Incorporate onTouchEvent and rotate the view
This is taken from android's OpenGL tutorial. It rotates the whole
view, not the cube. There will be glitches or overlaps in the rendered
cube as the order of rendering squares are not changed according to the
current rotation angle. To fix this, we should rotate the cube (or the
appropriate face of it) instead of the view.