Rotate individual faces according to touch/motion
Part 1 (works more or less okay): Axis of rotation
Find the slope of the line connecting the start and end points of the
motion. Use this slope value to deduce the axis of rotation.
Sometimes it rotates top face when I meant to rotate back face.

Part 2 (needs improvement): Which face to rotate
The face is chosen based on the proportionate position of touch for the
appropriate axis. This estimation works only for 3x3 cubes when the
phone is in portrait mode.