Highlight the touched point (read on)
The point is now projected (kind of) from a plane parallel to the front
face. This is why is looks a little off. We need to find the plane from
which user is watching (orthogonal to the eye vector) and tranlate the
touched point to that plane). Projecting that point to the origin should
give us the correct touch point. Or we might have to draw a line from
that point to a parallel plane behind the cube and figure out the
intersection.