Highlight the touched point on the cube
Touches happen on the screen which is a plane normal to the vector
from eye position to the origin. To get the touched point, first we
translate this plane parallel to the near plane of the frustrum. This
makes it easier to calculate the x and y positions in opengl
coordinates. These values are then translated back to the eye plane.

Imagine a line parallel to the eye vector that passes through this
touched point. Check if this line touches any of the visible planes of
the cube, namely front face, top face and the right face. If the point
of intersection of the line and a plane lies within the boundary
rectangle of the cube, we can assume the user touched that face.

Currently, we just highlight the touched point. Even though this method
shows improvement over the previous method (directly projecting near
plane to the front face), the estimated touch points are little off from
the actual points. Gotta solve this.