Count the moves taken for solving
The value is meant to be used during manual and automated solving.
It doesn't mean much during manual solving as user can invoke machine
solving and cancel at the last moment to reset the counter. Once we add
support for autodetecting solved cube during manual rotation, we should
find a way to use this value in a more meaningful way.