Fix rotation in case of 1xNxN cubes
The corner pieces of a skewed cube with height/width/depth of single
layer can have four or even five colored squares. Also, the center
pieces have two visible squres (white and yellow centers are on the same
piece in case of a 1xNxN cube. This means we need to perform an extra
set of rotations for such layers.

Rotate the opposite face in case the dimension along the axis of
rotation is one. By default, we rotate only the lower face; in case of
skewed cubes, the top face also needs to be updated.