Handle 90 degree rotations along assymmetric axes
An axis is assymmetric if the sizes along its orthogonal axes are not
equal. For instance, if the width of the cube is not equal to its
height, the cube isn't symmetric around z axis.

Individual faces cannot be rotated 90' along an assymmetric axis as it
will result in a disoriented cube. However, one should be able to rotate
the whole cube by 90' along any axis. To enable this, this patchset
updates the faces and recreates the layers according to the requested
axis and direction. Check rotateCubeX/Y/Z functions.

Contrary to our usual rotations, it is not enough to just update the
arrays here. The coordinates of all squares should also be changed. It
is accomplised using matrix multiplication provided by opengl. Now the
Cube class has a direct dependency on opengl.