Add logs for debugging crashes
There are two crases. One in RubiksCube.draw function when it tries to
read facelist.get(i) with value 3 from the list of size 3. Another was
an assertion in Cube.rotate() method where face being rotated is higher
than the number of layers on that axis. Both could be related to race
conditions between main thread and gl thread.